Enhance build configuration and library integration
- Updated .gitignore to include additional build artifacts and cache files for cleaner repository management.
- Added new submodule 'libhotcalls' to .gitmodules for improved library integration.
- Refactored CMakeLists.txt files across multiple components to standardize installation prefixes and include GNUInstallDirs for better path management.
- Enhanced package.sh script to streamline the build and packaging process, allowing for easier management of multiple components.
- Improved installation rules in libdune and libdunify to ensure proper handling of runtime and development files, facilitating better package management.

These changes improve the modularity, maintainability, and usability of the build system for the virtualization platform.

+++ src/core/vmpl.c
@@ -91,12 +91,12 @@ static struct vcpu_config *vcpu_config_alloc(struct vmpl_percpu *percpu)
     struct vcpu_config *config = malloc(sizeof(struct vcpu_config));
     memset(config, 0, sizeof(struct vcpu_config));
 
-    config->lstar = &__dune_syscall;
+    config->lstar = (uint64_t)&__dune_syscall;
     config->fs.base = percpu->kfs_base;
     config->gs.base = (uint64_t)percpu;
 
     config->tr.selector = GD_TSS;
-    config->tr.base = &percpu->tss;
+    config->tr.base = (uint64_t)&percpu->tss;
     config->tr.limit = sizeof(percpu->tss);
     config->tr.attrib = 0x0089; // refer to linux-svsm
